enum refactorizado
