+++ src/main/java/org/example/entity/Articulo.java
@@ -4,7 +4,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.*;
-import org.example.TipoLote;
+import org.example.enums.TipoLote;
 
 import java.time.LocalDateTime;
